# FamilyBenefits-api-openapi.yaml
openapi: 3.0.3
info:
  title: Family Benefits
  description: rest-api для приложения "Пособия для семьи с детьми"
  version: 1.0.0
  contact:
    email: smegovic@gmail.com
servers:
  - url: https://localhost
...
paths:
...
  /institution:
    post:
      tags:
        - institution
      summary: Добавляет новое учреждение
      description:
        Добавляет новое учреждение.
        Для администратора.
        Необходима авторизация.
      operationId: addInsitution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionAdd'
        required: true
      responses:
        '201':
          description: Учреждение добавлено
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '400':
          description: Учреждение с указанными данными уже существует
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    put:
      tags:
        - institution
      summary: Обновляет данные учреждения
      description:
        Обновляет данные учреждения.
        Для администратора.
        Необходима авторизация.
      operationId: updateInstitution
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstitutionUpdate'
        required: true
      responses:
        '201':
          description: Учреждение обновлено
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Учреждение с указанными данными не найдено
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /institution/{id}:
    delete:
      tags:
        - institution
      summary: Удаляет учреждение
      description:
        Удаляет учреждение.
        Для администратора.
        Необходима авторизация.
      operationId: deleteInstitution
      parameters:
        - name: id
          in: path
          description: ID учреждения
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Учреждение удалено
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Учреждение с указанными данными не найдено
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - institution
      summary: Возвращает информацию об учреждении
      description:
        Возвращает информацию об учреждении.
      operationId: getInstitution
      parameters:
        - name: id
          in: path
          description: ID учреждения
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные об учреждении получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionInfo'
        '404':
          description: Учреждение с указанным ID не найдено
          content: { }
      security: [ ]
...
  /institution/allpartial:
    get:
      tags:
        - institution
      summary: Возвращает множество учреждений, в которых нет пособий
      description:
        Возвращает множество учреждений, в которых нет пособий
        Для администратора.
        Необходима авторизация.
      operationId: getPartialInstitutions
      responses:
        '200':
          description: Множество учреждений получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstitutionInfo'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /institution/initdata:
    get:
      tags:
        - institution
      summary: Возвращает дополнительные данные для учреждения
      description:
        Возвращает дополнительные данные для добавления или обновления учреждения.
        Данные содержат в себе множество кратких информаций о городах и пособиях.
        Для администратора.
        Необходима авторизация.
      operationId: getInstitutionInitData
      responses:
        '200':
          description: Дополнительные данные учреждения получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstitutionInitData'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
...
  /criteriontype:
    post:
      tags:
        - criterion type
      summary: Добавляет новый тип критерия
      description:
        Добавляет новый тип критерия.
        Для администратора.
        Необходима авторизация.
      operationId: addCriterionType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionTypeAdd'
        required: true
      responses:
        '201':
          description: Тип критерия добавлен
          content: { }
        '400':
          description: Тип критерия с указанными данными уже существует
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    put:
      tags:
        - criterion type
      summary: Обновляет данные типа критерия
      description:
        Обновляет данные типа критерия.
        Для администратора.
        Необходима авторизация.
      operationId: updateCriterionType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriterionTypeUpdate'
        required: true
      responses:
        '201':
          description: Тип критерия обновлен
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Не найден тип критерия с указанными данными
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
  /criteriontype/{id}:
    delete:
      tags:
        - criterion type
      summary: Удаляет тип критерия
      description:
        Удаляет тип критерия.
        Для администратора.
        Необходима авторизация.
      operationId: deleteCriterionType
      parameters:
        - name: id
          in: path
          description: ID типа критерия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '201':
          description: Тип критерия удален
          content: { }
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
        '404':
          description: Тип критерия с указанным ID не найден
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
    get:
      tags:
        - criterion type
      summary: Возвращает информацию о типе критерия
      description:
        Возвращает информацию о типе критерия.
      operationId: getCriterionType
      parameters:
        - name: id
          in: path
          description: ID типа критерия
          schema:
            type: string
            example: td56387fhj
          required: true
      responses:
        '200':
          description: Данные о типе критерия получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CriterionTypeInfo'
        '404':
          description: Тип критерия с указанным ID не найден
          content: { }
      security: [ ]
...
  /criteriontype/allpartial:
    get:
      tags:
        - criterion type
      summary: Возвращает множество типов критерия, в которых нет критерий
      description:
        Возвращает множество типов критерия, в которых нет критерий.
        Для администратора.
        Необходима авторизация.
      operationId: getPartialCriterionTypes
      responses:
        '200':
          description: Множество типов критерия получено
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CriterionTypeInfo'
        '401':
          description: Клиент не авторизован или токен не действительный
          content: { }
        '403':
          description: Нет доступа
          content: { }
      security:
        - cookieAuth: [ ]
        - bearerAuth: [ ]
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CriterionTypeInfo:
      description: Информация о типе критерия
      properties:
        id:
          description: ID типа критерия
          type: string
          example: td56387fhj
        name:
          description: Название типа критерия
          type: string
          example: name
        info:
          description: Информация типа критерия
          type: string
          example: info
        shortCriterionSet:
          description: Множество кратких информаций о кртитериях типа критерия
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
    CriterionTypeAdd:
      description: Объект запроса для добавления типа критерия
      properties:
        name:
          description: Название типа критерия
          type: string
          example: name
        info:
          description: Информация типа критерия
          type: string
          example: info
    CriterionTypeUpdate:
      description: Объект запроса для обновления типа критерия
      properties:
        id:
          description: ID типа критерия
          type: string
          example: td56387fhj
        name:
          description: Название типа критерия
          type: string
          example: name
        info:
          description: Информация типа критерия
          type: string
          example: info
...
    InstitutionInfo:
      description: Информация об учреждении
      properties:
        id:
          description: ID учреждения
          type: string
          example: td56387fhj
        name:
          description: Название учреждения
          type: string
          example: name
        address:
          description: Адрес учреждения
          type: string
          example: address
        phone:
          description: Телефон учреждения
          type: string
          example: 8-888-888-88-88
        email:
          description: Электронная почта учржедения
          type: string
          example: example@mail.com
        schedule:
          description: График работы учреждения
          type: string
          example: График
        shortCity:
          $ref: '#/components/schemas/ObjectShortInfo'
        shortBenefitSet:
          description: Множество кратких информаций о пособиях учреждения
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
    InstitutionAdd:
      description: Объект запроса для добавления учреждения
      properties:
        name:
          description: Название учреждения
          type: string
          example: name
        address:
          description: Адрес учреждения
          type: string
          example: address
        phone:
          description: Телефон учреждения
          type: string
          example: 8-888-888-88-88
        email:
          description: Электронная почта учржедения
          type: string
          example: example@mail.com
        schedule:
          description: График работы учреждения
          type: string
          example: График
        idCity:
          description: ID города учреждения
          type: string
          example: td56387fhj
        idBenefitSet:
          description: Множество ID пособий учреждения
          type: array
          items:
            description: ID пособия
            type: string
            example: td56387fhj
    InstitutionUpdate:
      description: Объект запроса для обновления учреждения
      properties:
        id:
          description: ID учреждения
          type: string
          example: td56387fhj
        name:
          description: Название учреждения
          type: string
          example: name
        address:
          description: Адрес учреждения
          type: string
          example: address
        phone:
          description: Телефон учреждения
          type: string
          example: 8-888-888-88-88
        email:
          description: Электронная почта учржедения
          type: string
          example: example@mail.com
        schedule:
          description: График работы учреждения
          type: string
          example: График
        idCity:
          description: ID города учреждения
          type: string
          example: td56387fhj
        idBenefitSet:
          description: Множество ID пособий учреждения
          type: array
          items:
            description: ID пособия
            type: string
            example: td56387fhj
    InstitutionInitData:
      description: Дополнительные данные для добавления или обновления учреждения
        Содержат в себе множество кратких информаций о городах и пособиях
      properties:
        shortCitySet:
          description: Множество кратких информаций о городах
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
        shortBenefitSet:
          description: Множество кратких информаций о пособиях
          type: array
          items:
            $ref: '#/components/schemas/ObjectShortInfo'
...
    ObjectShortInfo:
      description: Краткая информация об объекте
      properties:
        idObject:
          description: ID объекта
          type: string
          example: td56387fhj
        nameObject:
          description: Название объекта
          type: string
          example: name